initial ideas, going to sleep for today
